+++ it/server/src/test/java/com/walmartlabs/concord/it/server/CheckpointsIT.java
@@ -47,6 +47,7 @@ public class CheckpointsIT extends AbstractServerIT {
 
         // start the process
 
+        CheckpointApi checkpointApi = new CheckpointApi(getApiClient());
         ProcessApi processApi = new ProcessApi(getApiClient());
         StartProcessResponse spr = start(payload);
         assertNotNull(spr.getInstanceId());
@@ -70,7 +71,7 @@ public class CheckpointsIT extends AbstractServerIT {
         // restore from TWO checkpoint
         String checkpointId = assertCheckpoint("TWO", processEvents);
 
-        ResumeProcessResponse resp = processApi.restore(pir.getInstanceId(), UUID.fromString(checkpointId));
+        ResumeProcessResponse resp = checkpointApi.restore(pir.getInstanceId(), UUID.fromString(checkpointId));
         assertNotNull(resp);
 
         waitForCompletion(processApi, spr.getInstanceId());
@@ -83,7 +84,7 @@ public class CheckpointsIT extends AbstractServerIT {
         // restore from ONE checkpoint
         checkpointId = assertCheckpoint("ONE", processEvents);
 
-        resp = processApi.restore(pir.getInstanceId(), UUID.fromString(checkpointId));
+        resp = checkpointApi.restore(pir.getInstanceId(), UUID.fromString(checkpointId));
         assertNotNull(resp);
 
         waitForCompletion(processApi, spr.getInstanceId());